+++ dynawo/sources/Modeler/Common/DYNSubModel.h
@@ -1257,7 +1257,7 @@ class SubModel {
    * @return number of calculated variables
    */
   inline unsigned int sizeCalculatedVar() const {
-    return calculatedVars_.size();
+    return static_cast<unsigned int>(calculatedVars_.size());
   }
 
   /**
@@ -1266,7 +1266,7 @@ class SubModel {
    * @return number of residual functions of init
    */
   inline unsigned int sizeFInit() const {
-    return fEquationInitIndex_.size();
+    return static_cast<unsigned int>(fEquationInitIndex_.size());
   }
 
   /**
@@ -1275,7 +1275,7 @@ class SubModel {
    * @return number of root functions of init
    */
   inline unsigned int sizeGInit() const {
-    return gEquationInitIndex_.size();
+    return static_cast<unsigned int>(gEquationInitIndex_.size());
   }
 
   /**
