+++ dynawo/sources/Modeler/DataInterface/PowSyblIIDM/DYNLoadInterfaceIIDM.cpp
@@ -67,9 +67,9 @@ LoadInterfaceIIDM::exportStateVariablesUnitComponent() {
     // should be removed once a solution has been found to propagate switches (de)connection
     // following component (de)connection (only Modelica models)
     if (connected && !getInitialConnected())
-      getVoltageLevelInterfaceInjector()->connectNode(loadIIDM_.getTerminal().getNodeBreakerView().getNode());
+      getVoltageLevelInterfaceInjector()->connectNode(static_cast<unsigned int>(loadIIDM_.getTerminal().getNodeBreakerView().getNode()));
     else if (!connected && getInitialConnected())
-      getVoltageLevelInterfaceInjector()->disconnectNode(loadIIDM_.getTerminal().getNodeBreakerView().getNode());
+      getVoltageLevelInterfaceInjector()->disconnectNode(static_cast<unsigned int>(loadIIDM_.getTerminal().getNodeBreakerView().getNode()));
   }
 
   if (connected)
