+++ dynawo/sources/Models/CPP/ModelNetwork/test/TestStaticVarCompensator.cpp
@@ -254,7 +254,7 @@ TEST(ModelsModelNetwork, ModelNetworkStaticVarCompensatorDiscreteVariables) {
   std::map<int, std::string> gEquationIndex;
   svc->setGequations(gEquationIndex);
   ASSERT_EQ(gEquationIndex.size(), nbG);
-  for (size_t i = 0; i < nbG; ++i) {
+  for (unsigned i = 0; i < nbG; ++i) {
     ASSERT_TRUE(gEquationIndex.find(i) != gEquationIndex.end());
   }
   delete[] zConnected;
